pyproject.toml
[project]
name = "ss-utils-bamboohr"
version = "0.1.0"
description = "BambooHR API"
requires-python = ">=3.12,<3.13"
authors = [{ name = "shern" }]
keywords = []
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "orjson>=3.11.7",
    "backoff>=2.2.1",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
]

[project.urls]
Repository = "https://github.com/shern2/ss-utils-bamboohr"
Issues = "https://github.com/shern2/ss-utils-bamboohr/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]


# [project.scripts]
# [package_name] = "[package_name].cli:app"


# --- Tooling Configuration ---
[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pyrefly>=0.46.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-vcr>=1.0.2",
    "python-dotenv>=1.2.1",
    "respx>=0.22.0",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 120
target-version = "py312"
fix = true
unsafe-fixes = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "F", # default Ruff rules
    "E", # pycodestyle Errors
    "W", # pycodestyle Warnings
    "I", # import sorting
    "UP", # modern syntax updates
    "B", # flake8 common pitfalls & best practices
    "Q", # flake8 consistent quotes
    "N", # naming conventions,
]
ignore = [
    "F401", # Unused import
    "E501", # line too long
    "B008", # function call in argument defaults
    "D400", # first line in docstring should end with period
    # "D413", # missing blank line after last section
    # "W293", # blank line contains whitespace
    # "W291", # trailing whitespace
]

fixable = ["ALL"]

extend-select = [
    "PLR2004", # magic value used in comparison
    "D101",    # missing docstring in public class
    "D102",    # missing docstring in public method
    "D103",    # missing docstring in public function
    "D201",    # no blank lines before function docstring
    "D202",    # no blank lines after function docstring
    "D215",    # section underline is over-indented
    "T201",    # 'print' found (should use logging)
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.per-file-ignores]
# Allow print statements in tests, examples, and notebooks
"tests/**/*.py" = ["T201", "PLR2004"]
"examples/**/*.py" = ["T201"]
"notebooks/**/*.ipynb" = ["T201", "F401"]
# camelCase field names mirror BambooHR API JSON keys exactly
"src/ss_utils_bamboohr/models.py" = ["N815"]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "pydantic.field_validator"]

[tool.pyrefly]
search-path = ["src", "."]

[tool.pytest.ini_options]
# Test discovery
testpaths = ["tests"]
pythonpath = ["src"]

# Execution options
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--asyncio-mode=auto",
    "--cov=.",
]

# Async test configuration
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Test markers
markers = [
    "unit: marks tests as unit tests (fast tests using mocks)",
    "integration: marks tests as integration tests (real API calls)",
    "e2e: marks tests as end-to-end workflow tests",
    "experimental: marks tests as experimental/WIP tests",
    "slow: marks tests as slow running (deselect with '-m \"not slow\"')",
    "asyncio: marks tests as async tests",
]
# Coverage configuration
[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "logs/*",
    "examples/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*.egg-info/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]


